add floating-fixed display mode using flag 32
Add "float_fix" mode which behaves like a hybrid of FIX and ALL modes,
based on the FLOAT mode on the TI-89.

Set DISP mode to FIX, and set Flag 32 to enable this mode. It behaves
like ALL mode with a limit to the max number of decimal places that will
be displayed. This is useful to avoid extra trailing zeros, but also
limit clutter an unwanted scientific notation that happens sometimes in
ALL mode.