android: add option to use haptic feedback API instead of vibrator
There's a standard View.performHapticFeedback API that can be used for
various normal events like key clicks. On some hardware, like my Pixel
3, this is wired up to different better-feeling hardware that doesn't
match any vibration duration.

Add a preferences button to use this mode instead of a normal vibration
duration. To avoid messing with the state file, store this as "-1" for
the keyVibration integer.