android: don't re-use main menu dialog
New in Android 10, all subsequent openings of the main menu have list
items which are only clickable underneath the text, not the full row.
It's fine on the first time that mainMenuDialog is displayed, and I have
no idea why.

Work around by removing the cached mainMenuDialog AlertDialog object and
re-building it every time the menu is shown. Also directly declare the
array of items rather than going through an ArrayList.