android: update to androidx appcompat libraries
This gives us updated UI styles in the various menus for "free" but it's
not pretty enough yet to be fully usable. A lot of things are smaller
and the margins aren't quite right. I probably need to update the
layouts, and maybe move things like AboutDialog to dedicated XML files.