common: use new/delete rather than malloc/realloc for phloat arrays
In BCD builds 'phloat' is a C++ class, and realloc on Phloat arrays
causes class-memaccess warnings "moving an object of non-trivially
copyable type `class Phloat`...". Using new[]/delete[] (or STL) is the
correct way to do this in C++. memcpy is similarly problematic, so
use assignment in a for loop to simluate realloc.